read folder as stack
not working:

ReaderPluginError: Tried to read ..\sub_stack with plugin napari, because it was associated with that file extension/because it is the only plugin capable of reading that path, but it gave an error. Try associating a different plugin or installing a different plugin for this kind of file.